+++ src/kard/transactions/types/__init__.py
@@ -26,6 +26,7 @@ if typing.TYPE_CHECKING:
     from .create_file_upload_request_body import CreateFileUploadRequestBody
     from .create_file_upload_url_response import CreateFileUploadUrlResponse
     from .direction_type import DirectionType
+    from .earned_rewards_range import EarnedRewardsRange
     from .file_upload_type import FileUploadType
     from .file_upload_url_attributes import FileUploadUrlAttributes
     from .file_upload_url_data import FileUploadUrlData
@@ -103,6 +104,7 @@ _dynamic_imports: typing.Dict[str, str] = {
     "CreateFileUploadRequestBody": ".create_file_upload_request_body",
     "CreateFileUploadUrlResponse": ".create_file_upload_url_response",
     "DirectionType": ".direction_type",
+    "EarnedRewardsRange": ".earned_rewards_range",
     "FileUploadType": ".file_upload_type",
     "FileUploadUrlAttributes": ".file_upload_url_attributes",
     "FileUploadUrlData": ".file_upload_url_data",
@@ -199,6 +201,7 @@ __all__ = [
     "CreateFileUploadRequestBody",
     "CreateFileUploadUrlResponse",
     "DirectionType",
+    "EarnedRewardsRange",
     "FileUploadType",
     "FileUploadUrlAttributes",
     "FileUploadUrlData",